# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "API Title",
        "version": "1.0"
    },
    "servers": [
        { "url": "http://localhost:8000" },
        { "url": "https://files-service.herokuapp.com" }
    ],
    "components": {
        "schemas": {
            "Image": {
                "type": "object",
                "properties": {
                    "filename": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    }
                }
            },
            "ResizedImage": {
                "type": "object",
                "properties": {
                    "originalname": {
                        "type": "string"
                    },
                    "original": {
                        "$ref": "#/components/schemas/Image"
                    },
                    "large": {
                        "$ref": "#/components/schemas/Image"
                    },
                    "medium": {
                        "$ref": "#/components/schemas/Image"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/Image"
                    }
                }
            },
            "File": {
                "type": "object",
                "properties": {
                    "fieldname": {
                        "type": "string"
                    },
                    "originalname": {
                        "type": "string"
                    },
                    "encoding": {
                        "type": "string"
                    },
                    "mimetype": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "size": {
                        "type": "number"
                    }
                }
            }
        }
    },
    "paths": {
        "/upload/files": {
            "post": {
                "description": "upload accept any type of files",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": ["files"],
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "userId": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/File"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/images": {
            "post": {
                "description": "image management",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "gallery": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/ResizedImage"
                                            }
                                        },
                                        "cover": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/ResizedImage"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/clear": {
            "delete": {
                "description": "file delete",
                "parameters": [
                    {
                        "in": "query",
                        "name": "directory",
                        "schema": {
                            "type": "string"
                        },
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/files/s3": {
            "post": {
                "description": "upload file to aws s3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": ["files"],
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "userId": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/File"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}